Fix label conversion in pipe.py

+++ notebooks/pipe.py
@@ -420,7 +420,7 @@ if args.train:
                     optimizer.zero_grad()
                     outputs = model(inputs)
                     labels_onehot = torch.zeros(labels.size(0), 2)
-                    labels_onehot.scatter_(1, labels.view(-1, 1), 1)
+                    labels_onehot.scatter_(1, labels.view(-1, 1).long(), 1)
                     labels_onehot = labels_onehot.to(outputs.device)  # Move labels tensor to the same device as outputs
                     loss = criterion(outputs, labels_onehot)
                     loss.backward()
